requirements.txt
apache-airflow~=2.7.3
apache-airflow-providers-cncf-kubernetes~=4.4.0
